Change lists, not tuples, in config settings
In the documentation, both `SOCIAL` and `LINKS` are described as a list
of tuples, but all examples show them as tuples of tuples. It makes the
most sense for them to be lists, so the examples should match the docs
going forward.

Similarly, `LOCALE` is described as a list, but the examples show it as
a tuple. Likewise, the examples should be lists, in line with the
description.