tasks.py.jinja2: add "new" task to create new content
Sometimes one wants to write, but does not have the energy to handle the
boilerplate and file management tasks. Add a `new` invoke task that
prompts for a title, proposes a slug based on that title, and creates
the appropriate file with its content based on the current date.

Defaults are set to using a flat `content/` path and `.rst` files.
Markdown is also supported, as well as a `content/<year>/` structure
commented out and ready to use.